move userReaction websocket handling into handlers file
