+++ webapp/src/index.tsx
@@ -30,6 +30,7 @@ import {
     handleUserVoiceOff,
     handleUserRaisedHand,
     handleUserUnraisedHand,
+    handleUserReaction,
 } from './websocket_handlers';
 
 import {
@@ -69,7 +70,6 @@ import reducer from './reducers';
 
 import {
     getPluginPath,
-    getPluginStaticPath,
     hasPermissionsToEnableCalls,
     getExpandedChannelID,
     getProfilesByIds,
@@ -186,38 +186,7 @@ export default class Plugin {
         });
 
         registry.registerWebSocketEventHandler(`custom_${pluginId}_user_reaction`, (ev) => {
-            // Note: reactions will not respond to displayname preferences, but they're only on screen for a short time
-            // anyway, so that's ok. (cf. other competitor's displayname doesn't update at all during an entire call).
-            const profiles = idToProfileInChannel(store.getState(), ev.broadcast.channel_id);
-            const displayName = getUserDisplayName(profiles[ev.data.userID]);
-            const reaction: Reaction = {
-                emoji: {
-                    name: ev.data.emoji_name,
-                    skin: ev.data.emoji_skin,
-                    unified: ev.data.emoji_unified,
-                },
-                timestamp: ev.data.timestamp,
-                user_id: ev.data.userID,
-                displayName,
-            };
-            store.dispatch({
-                type: VOICE_CHANNEL_USER_REACTION,
-                data: {
-                    channelID: ev.broadcast.channel_id,
-                    userID: ev.data.userID,
-                    reaction,
-                },
-            });
-            setTimeout(() => {
-                store.dispatch({
-                    type: VOICE_CHANNEL_USER_REACTION_TIMEOUT,
-                    data: {
-                        channelID: ev.broadcast.channel_id,
-                        userID: ev.data.userID,
-                        reaction,
-                    },
-                });
-            }, 10000);
+            handleUserReaction(store, ev);
         });
     }
